preprocessing Code and natural language
This script will generate txt file for each post with the project name appended at the front and the number of post appended at the end. This txt file has preprocessed natural language tokens and code tokens as @token@. The preprocessing for natural language has done separately from the code to maintain the code's semantics.